[DASH] Fix ZeroDiv for not timebased segment templates without given segment duration

+++ src/common/AdaptiveTree.cpp
@@ -105,7 +105,7 @@ namespace adaptive
     return false;
   }
 
-  uint32_t AdaptiveTree::estimate_segcount(uint32_t duration, uint32_t timescale)
+  uint32_t AdaptiveTree::estimate_segcount(uint64_t duration, uint32_t timescale)
   {
     duration /= timescale;
     return static_cast<uint32_t>((overallSeconds_ / duration)*1.01);
